requirements.txt
fastapi 
uvicorn 
httpx 
openai 
python-dotenv 
pydantic 
spacy 
transformers 
bertopic 
umap-learn 
hdbscan 
scikit-learn
torch